requirements.txt
Django==4.1.2
django-colorfield
python-barcode
Pillow